ensured numeric casting when operating on counts@x to avoid names being written out for these very long vectors.
note: found this when trying to explain why p2 objects are so much
larger than the starting matrices. The names was one explanation. The
other is the fact that the matrix is size scaled - the variety of the @x
values obviously reduces compression. Perhaps a better solution would be
to do cell size normalization on the fly (similar to how the variance
normalization is now scaled). The added beenfit is that we could then
keep only one copy of the matrix (now in misc$rawCounts) ... the
downside is that a lot of outside code depends on direct $count access,
and also performanc emay suffer